Fix Sonar issues

+++ common/src/system/thread/SleepUtil.h
@@ -11,7 +11,7 @@ namespace urchin {
         public:
             template<class T> static bool stepSleep(int, const T*);
 
-            static void sleepUs(unsigned int);
+            static void sleepUs(long);
 
         private:
             static constexpr int SLEEP_STEP_TIME_MS = 50;
